+++ Source/DREAM3DLib/SyntheticBuilderFilters/PackPrimaryPhases.h
@@ -181,7 +181,7 @@ class DREAM3DLib_EXPORT PackPrimaryPhases : public AbstractFilter
 
     unsigned long long int Seed;
 
-	int32_t firstPrimaryField;
+	size_t firstPrimaryField;
 
     float sizex;
     float sizey;
